packages/server/pyproject.toml
[project]
name = "ego-voice-agent-server"
version = "0.1.0"
description = "Voice agent app server for Echo Genesis Online"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "google-adk>=1.22.1",
  "uvicorn>=0.32.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
